requirements.txt
# 数据处理
pandas>=1.5.0
openpyxl>=3.0.0

# Web 框架
Flask>=2.0.0
Jinja2>=3.0.0

# 地图可视化
folium>=0.14.0

# API 调用
dashscope>=1.14.0
requests>=2.28.0

# 环境变量管理（用于安全存储 API Key）
python-dotenv>=1.0.0

# 测试
pytest>=7.0.0
